Move test out of the windows build.
Because it can still run on windows server and fail because it doesn't
have `chown`.
